Reworked the payments page, now transactions, separating in 3 pages: list, add payment, add transfer, add received
